NVar/Set get slices with time through sliceWithTime()
This means that NVarSet[item] no longer returns a NVar but a line from
the data set.

Additionally the date variable in NVarSet was removed. The time tags are
now acquired through the first variable in the set. This allows for
variables that already have values to be grouped into a set.